refac: 페이지네이션 컴포넌트 분리, 서치바 로직 추가
