+++ src/app/search/page.tsx
@@ -1,15 +1,19 @@
 "use client";
 
 import { SearchTrack } from "@/types/search"; // 타입스크립트
-import Image from "next/image";
+import { searchTracks } from "@/lib/spotifyToken"; // 토큰
+
 import { useSearchParams } from "next/navigation";
 import { useEffect, useState } from "react";
-import { searchTracks } from "@/lib/spotifyToken"; // 토큰
+
+import Pagination from "@/components/Pagination";
+import Image from "next/image";
 
 export default function SearchPage() {
   const searchParams = useSearchParams();
   const query = searchParams.get("query") || "";
   const [results, setResults] = useState<SearchTrack[]>([]);
+
   const [totalPages, setTotalPages] = useState(1);
   const [currentPage, setCurrentPage] = useState(1);
   const limit = 20;
@@ -38,7 +42,7 @@ export default function SearchPage() {
     }
   };
 
-  // 검색어가 바뀌면 실행
+  // query 바뀌면 실행
   useEffect(() => {
     if (query) {
       setResults([]);
@@ -52,68 +56,6 @@ export default function SearchPage() {
   const movePage = (page: number) => {
     setCurrentPage(page);
     fetchData(page);
-  };
-
-  const renderPageNumbers = () => {
-    const pageNumbers = [];
-    const startPage = Math.max(1, currentPage - Math.floor(pageRange / 2));
-    const endPage = Math.min(totalPages, startPage + pageRange - 1);
-
-    if (currentPage > 1) {
-      pageNumbers.push(
-        <button
-          key="prev"
-          onClick={() => movePage(Math.max(1, currentPage - pageRange))}
-          style={{
-            margin: "0 5px",
-            padding: "5px 10px",
-            backgroundColor: "#f0f0f0",
-            border: "1px solid #ddd",
-            cursor: "pointer"
-          }}
-        >
-          이전
-        </button>
-      );
-    }
-
-    for (let i = startPage; i <= endPage; i++) {
-      pageNumbers.push(
-        <button
-          key={i}
-          onClick={() => movePage(i)}
-          style={{
-            margin: "0 5px",
-            padding: "5px 10px",
-            backgroundColor: i === currentPage ? "#0079FF" : "#f0f0f0",
-            border: "1px solid #ddd",
-            cursor: "pointer"
-          }}
-        >
-          {i}
-        </button>
-      );
-    }
-
-    if (currentPage < totalPages) {
-      pageNumbers.push(
-        <button
-          key="next"
-          onClick={() => movePage(Math.min(totalPages, currentPage + pageRange))}
-          style={{
-            margin: "0 5px",
-            padding: "5px 10px",
-            backgroundColor: "#f0f0f0",
-            border: "1px solid #ddd",
-            cursor: "pointer"
-          }}
-        >
-          다음
-        </button>
-      );
-    }
-
-    return pageNumbers;
   };
 
   return (
@@ -130,7 +72,9 @@ export default function SearchPage() {
         ))}
       </ul>
 
-      <div style={{ marginTop: "20px" }}>{renderPageNumbers()}</div>
+      <div style={{ marginTop: "20px" }}>
+        <Pagination totalPages={totalPages} currentPage={currentPage} pageRange={pageRange} movePage={movePage} />
+      </div>
     </div>
   );
 }
